add dog weekly scheduling feature
As a manager, I need to be able to add new dogs to the system and
specify what slots they are signed up for.

Solution: a dog CRUD interface. Delete was not asked for yet, so leaving
off for now.

Note: I went with a series of "Day" booleans for ease of
querying, but it is possible that a serialized array of strings stored
in a text field could be better.

Note: I am testing at the web interface for now as it is fast when using
rack test. If necessary I will write tests at a lower level, but this
should get 100% coverage for now.

Note: I am bringing a simplified version of role_play abstraction that I wrote. I may add users later, but just imagine that I am using devise + roles.

Note: table_hash is also something cool I have done to make UI testing
nicer. It converts a table into an array of hashes, which makes the
comparison output pretty nice.

Note: added 'slim' as it is faster to type than erb and also because it
always generates correct html, unlike erb templating.

Note: I added a navigation bar at the top to get to various features. I
know there is going to be a different section to actually assign dogs to
trucks in the future, so lets do that now.