add ability to create day schedules
As a manager I need to be able to assign dogs to trucks for a specific
day, so that walkers can know their assignments.

Solution: added schedules CRUD routes and backend. I went with an
assignments text field because assignments is dynamic.

Note: the schedules index page is aggregating the assignments into just
counts for dogs and trucks.