add walker trip feature
As a dog walker, I need to be able to take a scheduled truck, mark dogs
as picked up, mark dogs as at park, mark dogs as leaving park, and mark dogs as dropped off.

Note: I wanted to make it easier for walkers to operate their app so
there is some complication on the server from that in trips around
knowing what to display to the walker.

Note: added pickup and dropoff resources to store when dog interactions.

Note: added status column to the dogs index table. The dog status logic goes in reverse chronological order of the trip as I
think that is going to give the correct status.